+++ reactor-ddb-om/src/test/java/com/ravenpack/aws/sample/it/TableCreationInsertAndMappingIT.java
@@ -1,9 +1,9 @@
 package com.ravenpack.aws.sample.it;
 
 import com.ravenpack.aws.reactor.Localstack;
-import com.ravenpack.aws.reactor.ReactorAWS;
 import com.ravenpack.aws.reactor.TestHelperDynamoDB;
 import com.ravenpack.aws.reactor.ddb.RxDynamo;
+import com.ravenpack.aws.reactor.ddb.RxDynamoImpl;
 import com.ravenpack.aws.sample.model.*;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.AfterAll;
@@ -29,7 +29,7 @@ class TableCreationInsertAndMappingIT
     private final TestHelperDynamoDB testHelperDynamoDB = new TestHelperDynamoDB(localstack);
 
     private  DynamoDbAsyncClient ddbClient = testHelperDynamoDB.getDdbAsyncClient();
-    private final RxDynamo rxDynamo = ReactorAWS.dynamo(ddbClient);
+    private final RxDynamo rxDynamo = new RxDynamoImpl(ddbClient);
 
     @AfterAll
     static void cleanup()