+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+2020-09-02
+* package split
 2020-07-07 
 * RxSQS delete is not "swallowing" errors
 * SQS operations operating on a single messages
@@ -5,3 +7,4 @@
     directly within RxSQS
 * DynamoDB object mapper - major refarctoring how model of data class is read
 * DynamoDB object mapper - support for tree like structures
+